Fixed the doctype entity problem relying on internet being available
The entity is now resolved locally when recognized. Moved the DTD file back to
the resources directory since it should be included in the .jar file. It is probably
visible enough now that it is saved in each map file. The DTD is now the one
by Olivier that was also added to the website.

The question is whether other people will have similar problems with such an
entity. I also wonder why setting setExpandEntityReferences to false is not enough
to make it not resolve the entities. Probably it means something else.