Add ability to rearange tilesets and rework the Tileset Manager.
Moved the  Save, Save As, Embed, Delete, and Edit buttons to the side of the
dialog. Added two buttons to the bottom for moving tile sets around. Added a
function in the Map class to swap tile sets and update the nessesary first
gids. Also added a column to the table to show the source of the tileset (or
"(Embedded)" if the set is embedded in the map file).