Fixed Layer resize undo and added TileRaster experimental class to new LayerResizeEdit
I fixed the undo functionality for layer resize. It should now work as expected with the current UI (couldn't test offset resize yet as the UI does not allow for it yet. I will implement this soon hopefully).

To support the resize UI, I created a new LayerResizeEdit which just captures the layer's data change required for a resize. This should be relatively small for smaller resizes; it avoids backing up the whole layer.

The new resize edit class uses an experimental TileRaster nested class. This class could be used throughout the project for storing tile data; which is currently a subfeature of MapLayer/TileLayer. However, some parts of the editor use TileLayer for things that are not too layer related, like the selection or brushes. For these, the TileRaster should do.
The idea is that every layer has a TileRaster, which does the actual work.

Currently TileRaster uses two dimensional arrays, but already leverages System.arraycopy for its copyRect operation. Using a one dimensional array might speed things up further and should be more memory fragmentation friendly.