Made parallax editing optional, fixed scroll wheel problems
Parallax editing is now optional. I replaced the viewport visibility toggle by a toggle action that would enable/disable not just the viewport display, but all parallax editing (this includes the parallax panel) and turn parallax rendering off in the MapView. This should be off by default so that new users won't be confused.

The scroll wheel issue is fixed by manually propagating the scroll wheel event to the mapView's scroll pane.