Added start of a selection tool to tiled.
Currently, a rubberband is implemented that only shows up when the marquee tool is selected when an object group layer is the current layer.

Under the hood, this thing uses a new ToolSemantics architecture. This basically moves tool code outside the editor, while allowing for multiple semantics for one and the same tool (selection might work differently for different layer types, for instance)